Refactor Class Balanced
